image and video have upload firebase

+++ src/main/java/com/example/social_network_fpt_be/cloudStorage/image/ImageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 
-public interface ImageRepository extends JpaRepository<Image, Integer>{
+public interface ImageRepository extends JpaRepository<Image, Long>{
 
 }
